Update Github workflow and terraform backend
The commit modifies the workflow name in .github/workflows/terraform.yml and adds the AWS S3 bucket configuration for terraform state in main.tf. The README.md also includes the prerequisites for the AWS environment. This update ensures the proper use of AWS S3 for Terraform state management and workflow naming convention.