Fix wrongly tagged workflow in GitHub Docker image config
The tag for the "Build Java with Maven" workflow in the configuration file responsible for building and pushing Docker images has been corrected. Previously, it had an incorrect '_removed' suffix, potentially preventing the GitHub Action from triggering correctly. This change ensures that the Docker image creation process is properly integrated with the Java build workflow.