Refactor ProductService and remove exception in ProductEventsHandler
Made sanitize method in ProductService static, updated usage in ProductCommandController, and removed an exception being thrown in the ProductEventshandler. Static method usage simplifies code and makes it more maintainable. The removed exception in ProductEventsHandler was unhandled and could result in unwanted application crashes. Tests were also expanded for the ProductService and ProductEventsHandler.